fix migration jobs

+++ internal/db/migrations/000002_create_jobs.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE jobs (
     id BIGSERIAL PRIMARY KEY,
     name TEXT NOT NULL,
-    organization_id INTEGER NOT NULL REFERENCES organizations(id)
+    organization_id BIGSERIAL NOT NULL REFERENCES organizations(id)
 );
